refactore fileNodeView
