+++ openmetadata-ui/src/main/resources/ui/src/components/BlockEditor/Extensions/File/AttachmentComponents/ImageAttachment.tsx
@@ -1,0 +1,90 @@
+/*
+ *  Copyright 2025 Collate.
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+import { NodeViewProps } from '@tiptap/react';
+import { Spin } from 'antd';
+import classNames from 'classnames';
+import React, { useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { ReactComponent as IconFormatImage } from '../../../../../assets/svg/ic-format-image.svg';
+import Loader from '../../../../common/Loader/Loader';
+
+const ImageAttachment = ({
+  node,
+  mediaSrc,
+  isMediaLoading,
+}: {
+  node: NodeViewProps['node'];
+  mediaSrc: string;
+  isMediaLoading: boolean;
+}) => {
+  const { url, alt, isUploading } = node.attrs;
+  const [imageError, setImageError] = useState<boolean>(false);
+  const [imageLoaded, setImageLoaded] = useState<boolean>(false);
+  const needsAuthentication = url?.includes('/api/v1/attachments/');
+  const { t } = useTranslation();
+
+  // Reset states when url changes
+  useEffect(() => {
+    setImageError(false);
+    setImageLoaded(false);
+  }, [url, mediaSrc]);
+
+  const handleImageError = () => {
+    setImageError(true);
+    setImageLoaded(false);
+  };
+
+  const handleImageLoad = () => {
+    setImageLoaded(true);
+    setImageError(false);
+  };
+
+  const showLoadingOverlay =
+    isUploading || (needsAuthentication && (!imageLoaded || isMediaLoading));
+  const displaySrc = needsAuthentication && !mediaSrc ? undefined : mediaSrc;
+
+  return (
+    <div className="image-wrapper">
+      <Spin
+        indicator={<Loader size="small" />}
+        spinning={showLoadingOverlay}
+        tip={isUploading ? t('label.uploading') : t('label.loading')}>
+        <div
+          className={classNames('image-container', {
+            'loading-state': showLoadingOverlay || imageError,
+          })}>
+          {(showLoadingOverlay || imageError) && (
+            <div className="loading-overlay">
+              <IconFormatImage width={40} />
+            </div>
+          )}
+          {displaySrc && (
+            <img
+              alt={alt ?? ''}
+              data-testid="uploaded-image-node"
+              src={displaySrc}
+              style={{
+                visibility: imageLoaded ? 'visible' : 'hidden',
+                display: 'block',
+              }}
+              onError={handleImageError}
+              onLoad={handleImageLoad}
+            />
+          )}
+        </div>
+      </Spin>
+    </div>
+  );
+};
+
+export default ImageAttachment;
