+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/de-de.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Identitätsquelle zum Annehmen der Rolle",
     "attention": "Achtung",
     "audience": "Zielgruppe",
+    "audio": "Audio",
     "august": "August",
     "auth-config-lowercase-plural": "Authentifizierungskonfigurationen",
     "auth-mechanism": "Authentifizierungsmechanismus",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Methode",
     "hyper-parameter-plural": "Hyperparameter",
     "icon-url": "Icon URL",
+    "image": "Bild",
     "image-repository": "Bildspeicher",
     "impact": "Auswirkung",
     "import": "Importieren",
@@ -1450,6 +1452,7 @@
     "upload": "Hochladen",
     "upload-csv-uppercase-file": "CSV-Datei hochladen",
     "upload-image": "Bild hochladen",
+    "uploading": "Hochladen",
     "uploading-file": "Datei wird hochgeladen",
     "upstream-depth": "Upstream-Tiefe",
     "url-lowercase": "URL",
@@ -1482,6 +1485,7 @@
     "version": "Version",
     "version-plural": "Versionen",
     "version-plural-history": "Versionshistorie",
+    "video": "Video",
     "view": "Ansicht",
     "view-all": "Alle anzeigen",
     "view-definition": "Ansichtsdefinition",
