+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/en-us.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Source Identity for Assume Role",
     "attention": "Attention",
     "audience": "Audience",
+    "audio": "Audio",
     "august": "August",
     "auth-config-lowercase-plural": "auth configs",
     "auth-mechanism": "Auth Mechanism",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "Hyper Parameters",
     "icon-url": "Icon URL",
+    "image": "Image",
     "image-repository": "Image Repository",
     "impact": "Impact",
     "import": "Import",
@@ -1450,6 +1452,7 @@
     "upload": "Upload",
     "upload-csv-uppercase-file": "Upload CSV file",
     "upload-image": "Upload Image",
+    "uploading": "uploading",
     "uploading-file": "Uploading file",
     "upstream-depth": "Upstream Depth",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "Version",
     "version-plural": "Versions",
     "version-plural-history": "Versions History",
+    "video": "Video",
     "view": "View",
     "view-all": "View All",
     "view-definition": "View Definition",
