+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/es-es.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Identidad de Origen para Asumir Rol",
     "attention": "Atención",
     "audience": "Público",
+    "audio": "Audio",
     "august": "Agosto",
     "auth-config-lowercase-plural": "Configuraciones de autenticación",
     "auth-mechanism": "Mecanismo de autenticación",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "Hiperparámetros",
     "icon-url": "Icono URL",
+    "image": "Imagen",
     "image-repository": "Repositorio de Imágenes",
     "impact": "Impacto",
     "import": "Importar",
@@ -1450,6 +1452,7 @@
     "upload": "Cargar",
     "upload-csv-uppercase-file": "Cargar archivo CSV",
     "upload-image": "Cargar Imagen",
+    "uploading": "Cargando",
     "uploading-file": "Subiendo archivo",
     "upstream-depth": "Profundidad aguas arriba",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "Versión",
     "version-plural": "Versiones",
     "version-plural-history": "Historial de versiones",
+    "video": "Video",
     "view": "Ver",
     "view-all": "Ver todo",
     "view-definition": "View Definition",
