+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/fr-fr.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Source d'Identité pour Assumer le Rôle",
     "attention": "Attention",
     "audience": "Audience",
+    "audio": "Audio",
     "august": "Août",
     "auth-config-lowercase-plural": "Configurations d'Authentification",
     "auth-mechanism": "Mécanisme d'Authentification",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "Hyperparamètres",
     "icon-url": "Icon URL",
+    "image": "Image",
     "image-repository": "Répertoire d'Images",
     "impact": "Impact",
     "import": "Importer",
@@ -1450,6 +1452,7 @@
     "upload": "Télécharger",
     "upload-csv-uppercase-file": "Télécharger le Fichier CSV",
     "upload-image": "Télécharger Image",
+    "uploading": "Téléchargement",
     "uploading-file": "Téléchargement du fichier",
     "upstream-depth": "Profondeur de l'Amont",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "Version",
     "version-plural": "Versions",
     "version-plural-history": "Historique des Versions",
+    "video": "Vidéo",
     "view": "Vue",
     "view-all": "Voir Tout",
     "view-definition": "Définition de la Vue",
