+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/gl-es.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Identidade da fonte para asumir o rol",
     "attention": "Atención",
     "audience": "Audiencia",
+    "audio": "Audio",
     "august": "Agosto",
     "auth-config-lowercase-plural": "configuracións de autenticación",
     "auth-mechanism": "Mecanismo de autenticación",
@@ -656,6 +657,7 @@
     "http-method": "Método HTTP",
     "hyper-parameter-plural": "Hiperparámetros",
     "icon-url": "URL do icono",
+    "image": "Imaxe",
     "image-repository": "Repositorio de imaxes",
     "impact": "Impacto",
     "import": "Importar",
@@ -1450,6 +1452,7 @@
     "upload": "Cargar",
     "upload-csv-uppercase-file": "Cargar ficheiro CSV",
     "upload-image": "Cargar imaxe",
+    "uploading": "Cargando",
     "uploading-file": "Cargando ficheiro",
     "upstream-depth": "Profundidade ascendente",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "Versión",
     "version-plural": "Versións",
     "version-plural-history": "Historial de versións",
+    "video": "Vídeo",
     "view": "Vista",
     "view-all": "Ver todo",
     "view-definition": "Ver definición",
