+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/he-he.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "זהות המקור להניח תפקיד",
     "attention": "תשומת לב",
     "audience": "קהל יעד",
+    "audio": "אודיו",
     "august": "אוגוסט",
     "auth-config-lowercase-plural": "הגדרות אימות",
     "auth-mechanism": "מנגנון אימות",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "הייפר פרמטרים",
     "icon-url": "כתובת URL לאייקון",
+    "image": "תמונה",
     "image-repository": "מאגר תמונות",
     "impact": "השפעה",
     "import": "ייבוא",
@@ -1450,6 +1452,7 @@
     "upload": "העלאה",
     "upload-csv-uppercase-file": "העלאת קובץ CSV",
     "upload-image": "העלאת תמונה",
+    "uploading": "העלאה",
     "uploading-file": "העלאת קובץ",
     "upstream-depth": "עומק למעלה",
     "url-lowercase": "כתובת URL",
@@ -1482,6 +1485,7 @@
     "version": "גרסה",
     "version-plural": "גרסאות",
     "version-plural-history": "היסטוריית גרסאות",
+    "video": "וידאו",
     "view": "תצוגה",
     "view-all": "הצג הכל",
     "view-definition": "הצג תיאור",
