+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ja-jp.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Source Identity for Assume Role",
     "attention": "Attention",
     "audience": "オーディエンス",
+    "audio": "音声",
     "august": "8月",
     "auth-config-lowercase-plural": "auth configs",
     "auth-mechanism": "Auth Mechanism",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "Hyper Parameters",
     "icon-url": "Icon URL",
+    "image": "画像",
     "image-repository": "Image Repository",
     "impact": "影響",
     "import": "インポート",
@@ -1450,6 +1452,7 @@
     "upload": "Upload",
     "upload-csv-uppercase-file": "CSVファイルをアップロード",
     "upload-image": "Upload Image",
+    "uploading": "アップロード中",
     "uploading-file": "ファイルをアップロード中",
     "upstream-depth": "Upstream Depth",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "バージョン",
     "version-plural": "バージョン",
     "version-plural-history": "バージョン履歴",
+    "video": "ビデオ",
     "view": "見る",
     "view-all": "全てを見る",
     "view-definition": "View Definition",
