+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ko-kr.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Assume Role의 Source Identity",
     "attention": "주의",
     "audience": "청중",
+    "audio": "오디오",
     "august": "8월",
     "auth-config-lowercase-plural": "인증 구성들",
     "auth-mechanism": "인증 메커니즘",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "하이퍼 파라미터들",
     "icon-url": "아이콘 URL",
+    "image": "이미지",
     "image-repository": "이미지 저장소",
     "impact": "Impact",
     "import": "가져오기",
@@ -1450,6 +1452,7 @@
     "upload": "업로드",
     "upload-csv-uppercase-file": "CSV 파일 업로드",
     "upload-image": "이미지 업로드",
+    "uploading": "업로드 중",
     "uploading-file": "Uploading file",
     "upstream-depth": "업스트림 깊이",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "버전",
     "version-plural": "버전들",
     "version-plural-history": "버전 기록들",
+    "video": "비디오",
     "view": "보기",
     "view-all": "전체 보기",
     "view-definition": "정의 보기",
