+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/mr-in.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "भूमिकेसाठी स्त्रोत ओळख",
     "attention": "लक्ष द्या",
     "audience": "प्रेक्षक",
+    "audio": "ऑडिओ",
     "august": "ऑगस्ट",
     "auth-config-lowercase-plural": "प्रमाणीकरण संरचना",
     "auth-mechanism": "प्रमाणीकरण यंत्रणा",
@@ -656,6 +657,7 @@
     "http-method": "HTTP पद्धत",
     "hyper-parameter-plural": "हायपर पॅरामीटर्स",
     "icon-url": "आयकॉन यूआरआय",
+    "image": "प्रतिमा",
     "image-repository": "प्रतिमा संग्रह",
     "impact": "प्रभाव",
     "import": "आयात करा",
@@ -1450,6 +1452,7 @@
     "upload": "अपलोड करा",
     "upload-csv-uppercase-file": "CSV फाइल अपलोड करा",
     "upload-image": "प्रतिमा अपलोड करा",
+    "uploading": "अपलोड करत आहे",
     "uploading-file": "फाइल अपलोड करत आहे",
     "upstream-depth": "अपस्ट्रीम खोली",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "आवृत्ती",
     "version-plural": "आवृत्त्या",
     "version-plural-history": "आवृत्ती इतिहास",
+    "video": "व्हिडिओ",
     "view": "पहा",
     "view-all": "सर्व पहा",
     "view-definition": "परिभाषा पहा",
