+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/nl-nl.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Source Identity for Assume Role",
     "attention": "Aandacht",
     "audience": "Doelgroep",
+    "audio": "Audio",
     "august": "augustus",
     "auth-config-lowercase-plural": "Auth-configuraties",
     "auth-mechanism": "Auth-mechanisme",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "Hyperparameters",
     "icon-url": "Pictogram URL",
+    "image": "Afbeelding",
     "image-repository": "Afbeeldingenrepository",
     "impact": "Impact",
     "import": "Importeren",
@@ -1450,6 +1452,7 @@
     "upload": "Uploaden",
     "upload-csv-uppercase-file": "CSV-bestand uploaden",
     "upload-image": "Afbeelding uploaden",
+    "uploading": "Uploaden",
     "uploading-file": "Bestand uploaden",
     "upstream-depth": "Upstreamdiepte",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "Versie",
     "version-plural": "Versies",
     "version-plural-history": "Versiegeschiedenis",
+    "video": "Video",
     "view": "View",
     "view-all": "Alle views",
     "view-definition": "Viewdefinitie",
