+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/pr-pr.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "هویت منبع برای فرض نقش",
     "attention": "توجه",
     "audience": "مخاطب",
+    "audio": "صوت",
     "august": "اوت",
     "auth-config-lowercase-plural": "پیکربندی‌های احراز هویت",
     "auth-mechanism": "مکانیزم احراز هویت",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "پارامترهای ابر",
     "icon-url": "آیکون URL",
+    "image": "Image",
     "image-repository": "مخزن تصویر",
     "impact": "تأثیر",
     "import": "وارد کردن",
@@ -1450,6 +1452,7 @@
     "upload": "بارگذاری",
     "upload-csv-uppercase-file": "بارگذاری فایل CSV",
     "upload-image": "بارگذاری تصویر",
+    "uploading": "بارگذاری",
     "uploading-file": "بارگذاری فایل",
     "upstream-depth": "عمق بالادست",
     "url-lowercase": "آدرس وب",
@@ -1482,6 +1485,7 @@
     "version": "نسخه",
     "version-plural": "نسخه‌ها",
     "version-plural-history": "تاریخچه نسخه‌ها",
+    "video": "ویدئو",
     "view": "مشاهده",
     "view-all": "مشاهده همه",
     "view-definition": "مشاهده تعریف",
