+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/pt-br.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Identidade Fonte para Assumir Papel",
     "attention": "Atenção",
     "audience": "Audiência",
+    "audio": "Áudio",
     "august": "Agosto",
     "auth-config-lowercase-plural": "configs de autenticação",
     "auth-mechanism": "Mecanismo de Autenticação",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "Hiperparâmetros",
     "icon-url": "URL do Ícone",
+    "image": "Imagem",
     "image-repository": "Repositório de Imagens",
     "impact": "Impacto",
     "import": "Importar",
@@ -1450,6 +1452,7 @@
     "upload": "Carregar",
     "upload-csv-uppercase-file": "Carregar arquivo CSV",
     "upload-image": "Carregar Imagem",
+    "uploading": "Carregando",
     "uploading-file": "Carregando arquivo",
     "upstream-depth": "Profundidade a Montante",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "Versão",
     "version-plural": "Versões",
     "version-plural-history": "Histórico de Versões",
+    "video": "Vídeo",
     "view": "Visualizar",
     "view-all": "Visualizar Tudo",
     "view-definition": "Definição de Visualização",
