+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ru-ru.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Идентификатор источника для принятой роли",
     "attention": "Внимание",
     "audience": "Аудитория",
+    "audio": "Аудио",
     "august": "Август",
     "auth-config-lowercase-plural": "конфигурация аутентификации",
     "auth-mechanism": "Механизм аутентификации",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "Hyper параметры",
     "icon-url": "Icon URL",
+    "image": "Изображение",
     "image-repository": "Репозиторий изображений",
     "impact": "Влияние",
     "import": "Импорт",
@@ -1450,6 +1452,7 @@
     "upload": "Upload",
     "upload-csv-uppercase-file": "Загрузить CSV-файл",
     "upload-image": "Upload Image",
+    "uploading": "Загрузка",
     "uploading-file": "Загрузка файла",
     "upstream-depth": "Восходящая линия",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "Версия",
     "version-plural": "Версии",
     "version-plural-history": "История версий",
+    "video": "Видео",
     "view": "Представление",
     "view-all": "Посмотреть все",
     "view-definition": "Посмотреть определение",
