+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/th-th.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "เอกลักษณ์ต้นทางสำหรับการถือบทบาท",
     "attention": "ความสนใจ",
     "audience": "ผู้ฟัง",
+    "audio": "เสียง",
     "august": "สิงหาคม",
     "auth-config-lowercase-plural": "การกำหนดค่าการรับรอง",
     "auth-mechanism": "กลไกการรับรอง",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "พารามิเตอร์ไฮเปอร์",
     "icon-url": "URL ไอคอน",
+    "image": "ภาพ",
     "image-repository": "ที่เก็บภาพ",
     "impact": "ผลกระทบ",
     "import": "นำเข้า",
@@ -1450,6 +1452,7 @@
     "upload": "อัปโหลด",
     "upload-csv-uppercase-file": "อัปโหลดไฟล์ CSV",
     "upload-image": "อัปโหลดภาพ",
+    "uploading": "กำลังอัปโหลด",
     "uploading-file": "อัปโหลดไฟล์",
     "upstream-depth": "ความลึกทางด้านขาเข้า",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "เวอร์ชัน",
     "version-plural": "เวอร์ชันหลายรายการ",
     "version-plural-history": "ประวัติเวอร์ชัน",
+    "video": "วิดีโอ",
     "view": "ดู",
     "view-all": "ดูทั้งหมด",
     "view-definition": "ดูคำจำกัดความ",
