+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/zh-cn.json
@@ -116,6 +116,7 @@
     "assume-role-source-identity": "Source Identity for Assume Role",
     "attention": "注意",
     "audience": "受众",
+    "audio": "音频",
     "august": "八月",
     "auth-config-lowercase-plural": "鉴权配置",
     "auth-mechanism": "鉴权机制",
@@ -656,6 +657,7 @@
     "http-method": "HTTP Method",
     "hyper-parameter-plural": "超参数",
     "icon-url": "Icon URL",
+    "image": "图片",
     "image-repository": "图像仓库",
     "impact": "影响",
     "import": "导入",
@@ -1450,6 +1452,7 @@
     "upload": "上传",
     "upload-csv-uppercase-file": "上传 CSV 文件",
     "upload-image": "上传图片",
+    "uploading": "上传中",
     "uploading-file": "上传文件",
     "upstream-depth": "上游深度",
     "url-lowercase": "url",
@@ -1482,6 +1485,7 @@
     "version": "版本",
     "version-plural": "版本",
     "version-plural-history": "版本历史",
+    "video": "视频",
     "view": "查看",
     "view-all": "查看全部",
     "view-definition": "查看定义",
